migrations doen het nu

+++ database/migrations/2022_03_17_124857_create_bankstatements_table.php
@@ -15,7 +15,7 @@ class CreateBankstatementsTable extends Migration
     {
         Schema::create('bankstatements', function (Blueprint $table) {
             $table->id();
-            $table->decimal('amount');
+            $table->double('amount');
             $table->string('name');
             $table->boolean('type');
             $table->dateTime('date');
